# docs/swagger.yaml
definitions:
...
  req.ActSearchReq:
    properties:
      detailTime:
        properties:
          end_time:
            type: string
          start_time:
            type: string
        type: object
      host:
        items:
          type: string
        type: array
      if_register:
        type: string
      location:
        items:
          type: string
        type: array
      type:
        items:
          type: string
        type: array
    type: object
  req.CreateActReq:
    properties:
      introduce:
        type: string
      labelform:
        properties:
          activeForm:
            type: string
          endTime:
            type: string
          holderType:
            type: string
          if_register:
            type: string
          position:
            type: string
          register_method:
            type: string
          signer:
            items:
              properties:
                name:
                  type: string
                studentid:
                  type: string
              type: object
            type: array
          startTime:
            type: string
          type:
            type: string
        type: object
      showImg:
        items:
          type: string
        type: array
      studentid:
        type: string
      title:
        type: string
    type: object
  req.CreateCommentReq:
    properties:
      content:
        type: string
      parent_id:
        type: string
      studentid:
        type: string
      subject:
        type: string
    type: object
...
  req.DeleteCommentReq:
    properties:
      studentid:
        type: string
      target_id:
        type: string
    type: object
  req.FindActByDateReq:
    properties:
      date:
        type: string
    type: object
  req.FindActByNameReq:
    properties:
      name:
        type: string
    type: object
...
  resp.CommentResp:
    properties:
      bid:
        type: string
      commented_pos:
        type: string
      commented_time:
        type: string
      content:
        type: string
      creator:
        properties:
          avatar:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
      likeNum:
        type: integer
      reply:
        items:
          properties:
            bid:
              type: string
            reply_content:
              type: string
            reply_creator:
              properties:
                avatar:
                  type: string
                studentid:
                  type: string
                username:
                  type: string
              type: object
            reply_pos:
              type: string
            reply_time:
              type: string
          type: object
        type: array
      replyNum:
        type: integer
    type: object
  resp.CreateActivityResp:
    properties:
      bid:
        type: string
      if_register:
        type: string
      introduce:
        type: string
      isChecking:
        type: string
      position:
        type: string
      showImg:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      userInfo:
        properties:
          avatar:
            type: string
          school:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
    type: object
...
  resp.ListActivitiesResp:
    properties:
      bid:
        type: string
      collectNum:
        type: integer
      commentNum:
        type: integer
      detailTime:
        properties:
          endTime:
            type: string
          startTime:
            type: string
        type: object
      holderType:
        type: string
      if_register:
        type: string
      introduce:
        type: string
      isCollect:
        type: string
      isLike:
        type: string
      likeNum:
        type: integer
      position:
        type: string
      showImg:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      userInfo:
        properties:
          avatar:
            type: string
          school:
            type: string
          studentid:
            type: string
          username:
            type: string
        type: object
    type: object
...
  resp.Resp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: 校灵通 API 文档
  title: EventGlide API
paths:
  /act/all/{id}:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.ListActivitiesResp'
              type: object
      summary: 列出所有活动
      tags:
      - Activity
  /act/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/req.CreateActReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.CreateActivityResp'
              type: object
      summary: 创建活动
      tags:
      - Activity
  /act/date:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 日期查找
        in: body
        name: date
        required: true
        schema:
          $ref: '#/definitions/req.FindActByDateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.ListActivitiesResp'
              type: object
      summary: 通过日期查找活动
      tags:
      - Activity
  /act/draft:
    post:
      consumes:
      - application/json
      description: not finished
      parameters:
      - description: 活动草稿
        in: body
        name: draft
        required: true
        schema:
          $ref: '#/definitions/req.CreateActReq'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.CreateActivityResp'
              type: object
      summary: 创建活动草稿
      tags:
      - Activity
  /act/load:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.CreateActivityResp'
              type: object
      summary: 加载活动草稿
      tags:
      - Activity
  /act/name:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 活动名称
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/req.FindActByNameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.ListActivitiesResp'
                  type: array
              type: object
      summary: 通过名称查找活动
      tags:
      - Activity
  /act/own/{id}:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建者id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.ListActivitiesResp'
              type: object
      summary: 通过创建者id查找活动
      tags:
      - Activity
  /act/search:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 搜索条件
        in: body
        name: actSearchReq
        required: true
        schema:
          $ref: '#/definitions/req.ActSearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.ListActivitiesResp'
              type: object
      summary: 通过搜索条件查找活动
      tags:
      - Activity
  /comment/answer:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 回复
        in: body
        name: CommentReq
        required: true
        schema:
          $ref: '#/definitions/req.CreateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.CommentResp'
              type: object
      summary: 回复评论
      tags:
      - Comment
  /comment/create:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论
        in: body
        name: CommentReq
        required: true
        schema:
          $ref: '#/definitions/req.CreateCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  $ref: '#/definitions/resp.CommentResp'
              type: object
      summary: 创建评论
      tags:
      - Comment
  /comment/delete:
    post:
      parameters:
      - description: 学号
        in: formData
        name: sid
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除评论
        in: body
        name: DeleteCommentReq
        required: true
        schema:
          $ref: '#/definitions/req.DeleteCommentReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Resp'
      summary: 删除评论
      tags:
      - Comment
  /comment/load/{id}:
    get:
      parameters:
      - description: 目标id
        in: path
        name: id
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resp.CommentResp'
                  type: array
              type: object
      summary: 加载评论
      tags:
      - Comment
...
swagger: "2.0"